requirements.txt
flask==3.1.1
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

